# openapi/cms-api.yaml
openapi: 3.1.0
info:
  title: CMS Collection Management System API
  version: 1.0.0
  description: |
    Multi-tenant Collection Management System API.
    
    ## Authentication
    All endpoints (except /health and /ready) require authentication via Bearer token.
    
    ## Multi-Tenancy
    All requests must include tenant context. In development, use the `x-tenant-id` header.
    In production, tenant is derived from the JWT token.
    
    ## Request Tracing
    Include `x-request-id` and `x-correlation-id` headers for distributed tracing.
  contact:
    name: CMS Team
  license:
    name: Proprietary
servers:
  - url: http://localhost:3001
    description: Local development server
  - url: https://api.cms.example.com
    description: Production server
...
paths:
...
  /api/v1/integration/ingest:
    post:
      operationId: ingestData
      summary: Main data ingestion endpoint (FILE/MANUAL/API)
      description: |
        Accepts data via FILE (multipart), MANUAL (JSON), or API (JSON) modes.
        Implements idempotency via (tenant_id, source_system_id, external_ref) + payload_hash.
      tags:
        - Integration - Ingest
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: CSV or XLSX file
                dataset:
                  type: string
                  enum: [PARTY, CONTRACT, CHARGE]
                source_system_code:
                  type: string
          application/json:
            schema:
              $ref: '#/components/schemas/IngestRequest'
      responses:
        '200':
          description: Ingestion successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestResponse'
        '207':
          description: Partial success - some records failed
        '422':
          description: Ingestion failed
  /api/v1/integration/ingest/method:
    get:
      operationId: getIntegrationMethod
      summary: Get configured integration method for a dataset
      tags:
        - Integration - Ingest
      parameters:
        - name: dataset
          in: query
          required: true
          schema:
            type: string
            enum: [PARTY, CONTRACT, CHARGE]
      responses:
        '200':
          description: Integration method for dataset
  /api/v1/integration/ingest/webhook-info:
    get:
      operationId: getWebhookInfo
      summary: Get webhook endpoint info and sample payload
      tags:
        - Integration - Ingest
      parameters:
        - name: dataset
          in: query
          schema:
            type: string
        - name: source_system_code
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Webhook info with sample payload

  # ========== EPIC 5: Integration - Runs ==========
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token with tenant and user claims
...
  schemas:
...
    IngestRequest:
      type: object
      required:
        - dataset
        - source_system_code
        - data
      properties:
        dataset:
          type: string
          enum: [PARTY, CONTRACT, CHARGE]
        source_system_code:
          type: string
        mode:
          type: string
          enum: [FILE, MANUAL, API]
        data:
          type: array
          items:
            type: object
        mapping:
          type: object
    IngestResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            run_id:
              type: string
              format: uuid
            status:
              type: string
              enum: [success, partial, failed]
            stats:
              type: object
              properties:
                totalReceived:
                  type: integer
                totalProcessed:
                  type: integer
                totalInserted:
                  type: integer
                totalUpdated:
                  type: integer
                totalSkipped:
                  type: integer
                totalFailed:
                  type: integer
            dq_issues_count:
              type: integer
        errors:
          type: array
          items:
            type: object
            properties:
              index:
                type: integer
              externalRef:
                type: string
              error:
                type: string
...
security:
  - bearerAuth: []